+++ babel.config.js
@@ -1,3 +1,13 @@
 module.exports = {
-  presets: [`@babel/preset-react`],
+  presets: [
+    `@babel/preset-react`,
+    [
+      `@babel/preset-env`,
+      {
+        targets: {
+          node: `current`,
+        },
+      },
+    ],
+  ],
 };
